+++ app/src/main/java/com/example/bearvbull/ui/components/ProfileScreenComponents.kt
@@ -246,7 +246,7 @@ fun UserNameWithEditButton(
 }
 
 @Composable
-fun EloWithRankImage(userEloScore: Int) {
+fun EloWithRankImage(userEloScore: Long) {
     val eloRank = getEloRank(userEloScore)
     Column(
         horizontalAlignment = Alignment.CenterHorizontally,
